move files to SearchLevelSelector, bug fixes
